travis: Add more build steps
Do install as root onto the system root, to ensure that that works, too.
Specifying a destdir is also valuable, however, as it's what packagers
do, so keep both - the install itself only takes second of a 20 minute
build process.

When the build itself is done and verified, try to pull new changes from
the submodules, rebuild, and reinstall. This won't pull new perl deps,
and should only rebuild changed sources, so it should be quick, too.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>