fixed racecondition in debian init script
stoping naemon before we wait for the process to fully stop sometimes
leads to the situation where the pid file is already removed but the
process still exists, so we read the pidfile before we do the stop.
Also let start-stop-daemon wait for the childs to exit, otherwise
child worker are still running sometimes and keep open filehandles which
prevent the pkg from a clean uninstall.

Signed-off-by: Sven Nierlein <Sven.Nierlein@consol.de>