move libnaemon.so into libnaemon package
it makes no sense to have the libnaemon.so file in the tools package. So do
proper packaging and put it into a seperate sub package. Also, since libnaemon.so
is required for building NEB modules, it makes no sense to require the tools
package just because of the libnaemon.so.

Signed-off-by: Sven Nierlein <Sven.Nierlein@consol.de>